pyproject.toml
[project]
name = "data-quality"
version = "0.1.0"
description = ""
authors = [
    {name = "Renan Domingos",email = "rna.domingos@hotmail.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.12"
dependencies = [
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-mermaid2-plugin (>=1.2.1,<2.0.0)",
    "mkdocs-material (>=9.6.14,<10.0.0)",
    "mkdocstrings[crystal,python] (>=0.29.1,<0.30.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pandera (>=0.24.0,<0.25.0)",
    "duckdb (>=1.3.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
format = """
isort .
black .
"""
kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
run = """
python3 app/main.py
"""
doc = "mkdocs serve"